+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Features
 
 - Show output file name for failing tests when using -o option
+- Add -O option to specify output directory for tap files
 
 ## [1.9.0] 2025-11-14